feat: put an helper function for creating a json object in jsonutils
