+++ app/src/main/java/com/android/bookswap/data/source/api/ChatGPTApiService.kt
@@ -3,9 +3,9 @@ package com.android.bookswap.data.source.api
 import android.content.Context
 import android.util.Log
 import com.android.bookswap.BuildConfig
+import com.android.bookswap.utils.createJsonObjectRequestOpenAI
+import com.android.volley.Request
 import com.android.volley.RequestQueue
-import com.android.volley.Response
-import com.android.volley.toolbox.JsonObjectRequest
 import com.android.volley.toolbox.Volley
 import org.json.JSONArray
 import org.json.JSONObject
@@ -102,53 +102,9 @@ class ChatGPTApiService(context: Context, mode: Boolean = true) : ApiService {
     Log.i("ChatGPTApiService", "Request to openAI API with: $jsonBody")
 
     // Create the request and listen for the response
-    val jsonObjectRequest =
-        object :
-            JsonObjectRequest(
-                Method.POST,
-                apiUrl,
-                jsonBody,
-                Response.Listener { response ->
-                  try {
-                    val choicesArray = response.getJSONArray("choices")
-                    val messageContent =
-                        choicesArray.getJSONObject(0).getJSONObject("message").getString("content")
-
-                    onSuccess(messageContent)
-                  } catch (e: Exception) {
-                    Log.e("ChatGPTApiService", "Error: ${e.localizedMessage}")
-                    onError("Parsing error: ${e.localizedMessage}")
-                  }
-                },
-                Response.ErrorListener { error ->
-                  Log.e("ChatGPTApiService", "Full error details: ", error)
-
-                  if (error.networkResponse != null) {
-                    val statusCode = error.networkResponse.statusCode
-                    val responseBody =
-                        error.networkResponse.data?.let { String(it) } ?: "No response body"
-                    Log.e(
-                        "ChatGPTApiService",
-                        "Status Code: $statusCode, Response Body: $responseBody")
-                  }
-
-                  val errorMessage = error.localizedMessage ?: "Unknown error"
-                  Log.e("ChatGPTApiService", "Error message: $errorMessage")
-                  onError("Request error: $errorMessage")
-                }) {
-          // Add the API key and content type to the request headers
-          override fun getHeaders(): MutableMap<String, String> {
-            return hashMapOf(
-                "Authorization" to "Bearer $apiKey", "Content-Type" to "application/json")
-          }
-          // Override the default timeout (in milliseconds)
-          override fun getRetryPolicy(): com.android.volley.RetryPolicy {
-            return com.android.volley.DefaultRetryPolicy(
-                30000, // Timeout in milliseconds (e.g., 30 seconds)
-                com.android.volley.DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
-                com.android.volley.DefaultRetryPolicy.DEFAULT_BACKOFF_MULT)
-          }
-        }
-    requestQueue.add(jsonObjectRequest)
+    val jsonObjectRequestOpenAI =
+        createJsonObjectRequestOpenAI(
+            Request.Method.POST, apiUrl, jsonBody, apiKey, onSuccess, onError)
+    requestQueue.add(jsonObjectRequestOpenAI)
   }
 }
