+++ app/src/main/java/com/android/bookswap/utils/JSONUtils.kt
@@ -1,5 +1,8 @@
 package com.android.bookswap.utils
 
+import android.util.Log
+import com.android.volley.Response
+import com.android.volley.toolbox.JsonObjectRequest
 import org.json.JSONArray
 import org.json.JSONObject
 
@@ -38,4 +41,67 @@ fun JSONObject.getStringOrNull(name: String): String? {
  */
 fun JSONArray.getStringOrNull(index: Int): String? {
   return if (!this.isNull(index)) this.getString(index) else null
+}
+
+/**
+ * Create a JsonObjectRequest for the OpenAI ChatGPT API.
+ *
+ * @param method The HTTP method to use.
+ * @param url The URL to send the request to.
+ * @param body The JSON body of the request.
+ * @param apiKey The API key to use for authorization.
+ * @param onSuccess The success callback function.
+ * @param onError The error callback function.
+ * @return The JsonObjectRequest object.
+ */
+fun createJsonObjectRequestOpenAI(
+    method: Int,
+    url: String,
+    body: JSONObject,
+    apiKey: String,
+    onSuccess: (String) -> Unit,
+    onError: (String) -> Unit
+): JsonObjectRequest {
+  return object :
+      JsonObjectRequest(
+          method,
+          url,
+          body,
+          Response.Listener { response ->
+            try {
+              val choicesArray = response.getJSONArray("choices")
+              val messageContent =
+                  choicesArray.getJSONObject(0).getJSONObject("message").getString("content")
+
+              onSuccess(messageContent)
+            } catch (e: Exception) {
+              Log.e("ChatGPTApiService", "Error: ${e.localizedMessage}")
+              onError("Parsing error: ${e.localizedMessage}")
+            }
+          },
+          Response.ErrorListener { error ->
+            Log.e("ChatGPTApiService", "Full error details: ", error)
+
+            if (error.networkResponse != null) {
+              val statusCode = error.networkResponse.statusCode
+              val responseBody =
+                  error.networkResponse.data?.let { String(it) } ?: "No response body"
+              Log.e("ChatGPTApiService", "Status Code: $statusCode, Response Body: $responseBody")
+            }
+
+            val errorMessage = error.localizedMessage ?: "Unknown error"
+            Log.e("ChatGPTApiService", "Error message: $errorMessage")
+            onError("Request error: $errorMessage")
+          }) {
+    override fun getHeaders(): MutableMap<String, String> {
+      return hashMapOf("Authorization" to "Bearer $apiKey", "Content-Type" to "application/json")
+    }
+
+    override fun getRetryPolicy(): com.android.volley.RetryPolicy {
+      return com.android.volley.DefaultRetryPolicy(
+          30000, // Timeout in milliseconds (e.g., 30 seconds)
+          com.android.volley.DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
+          com.android.volley.DefaultRetryPolicy.DEFAULT_BACKOFF_MULT)
+    }
+  }
 }
